GUI beetje opgeschoond, formules int -> double

+++ src/com/awesome/excelpp/math/Add.java
@@ -5,8 +5,8 @@ package com.awesome.excelpp.math;
  *
  */
 public class Add extends Formula {
-	public int getValue(int ... numbers) {
-		int sum = 0;
+	public double getValue(double ... numbers) {
+		double sum = 0;
 		for (int i = 0; i < numbers.length; i++)
 	        sum += numbers [i];
 		return sum;
