+++ src/com/awesome/excelpp/math/Average.java
@@ -1,8 +1,8 @@
 package com.awesome.excelpp.math;
 
 public class Average extends Formula {
-	public int getValue(int...numbers) {
-		int sum = 0;
+	public double getValue(double...numbers) {
+		double sum = 0;
 		int totalNumbers = 0;
 		for(int i=0; i < numbers.length; i++) {
 			sum += numbers[i];
